+++ JSONDb/IDataset.cs
@@ -7,9 +7,9 @@ namespace JSONDb
     public interface IDataset<E>
     {
         IEnumerable<E> GetAll();
-        object Find(int id);
+        object Find(long id);
         void Add(E value);
         void Update(E value);
-        void Remove(int id);
+        void Remove(long id);
     }
 }
